+++ src/index.ts
@@ -24,6 +24,7 @@ import {
   writeJson,
 } from 'fs-extra';
 import { createProxyMiddleware } from 'http-proxy-middleware';
+import { Socket } from 'net';
 import nodemon from 'nodemon';
 import { EOL, hostname } from 'os';
 import * as path from 'path';
@@ -139,8 +140,14 @@ class DevServer {
             alias: 'x',
             description: 'Do not build and install the adapter before starting.',
           },
+          doNotWatch: {
+            type: 'string',
+            alias: 'w',
+            description:
+              'Do not watch the given files or directories for changes (provide paths relative to the adapter base directory.',
+          },
         },
-        async (args) => await this.watch(!args.noStart, !!args.noInstall),
+        async (args) => await this.watch(!args.noStart, !!args.noInstall, args.doNotWatch),
       )
       .command(
         ['debug [profile]', 'd'],
@@ -376,7 +383,14 @@ class DevServer {
     await this.startServer();
   }
 
-  async watch(startAdapter: boolean, noInstall: boolean): Promise<void> {
+  async watch(startAdapter: boolean, noInstall: boolean, doNotWatch: string | string[] | undefined): Promise<void> {
+    let doNotWatchArr: string[] = [];
+    if (typeof doNotWatch === 'string') {
+      doNotWatchArr.push(doNotWatch);
+    } else if (Array.isArray(doNotWatch)) {
+      doNotWatchArr = doNotWatch;
+    }
+
     await this.checkSetup();
     if (!noInstall) {
       await this.buildLocalAdapter();
@@ -384,12 +398,12 @@ class DevServer {
     }
     if (this.isJSController()) {
       // this watches actually js-controller
-      await this.startAdapterWatch(startAdapter);
+      await this.startAdapterWatch(startAdapter, doNotWatchArr);
       await this.startServer();
     } else {
       await this.startJsController();
       await this.startServer();
-      await this.startAdapterWatch(startAdapter);
+      await this.startAdapterWatch(startAdapter, doNotWatchArr);
     }
   }
 
@@ -433,7 +447,7 @@ class DevServer {
       const dependencies = pkg.dependencies;
       return [
         name,
-        `http://localhost:${infos.adminPort}`,
+        `http://127.0.0.1:${infos.adminPort}`,
         dependencies['iobroker.js-controller'],
         dependencies['iobroker.admin'],
       ];
@@ -485,16 +499,71 @@ class DevServer {
     return existsSync(jsControllerDir);
   }
 
+  checkPort(port: number, host = '127.0.0.1', timeout = 1000): Promise<void> {
+    return new Promise((resolve, reject) => {
+      const socket = new Socket();
+
+      const onError = (): void => {
+        socket.destroy();
+        reject();
+      };
+
+      socket.setTimeout(timeout);
+      socket.once('error', onError);
+      socket.once('timeout', onError);
+
+      socket.connect(port, host, () => {
+        socket.end();
+        resolve();
+      });
+    });
+  }
+
+  async waitForPort(port: number, offset = 0): Promise<boolean> {
+    port = this.getPort(port, offset);
+    this.log.debug(`Waiting for port ${port} to be available...`);
+    let tries = 0;
+    while (true) {
+      try {
+        await this.checkPort(port);
+        this.log.debug(`Port ${port} is available...`);
+        return true;
+      } catch {
+        if (tries++ > 30) {
+          this.log.error(`Port ${port} is not available after 30 seconds.`);
+          return false;
+        }
+        await this.delay(1000);
+      }
+    }
+  }
+
+  async waitForJsController(): Promise<void> {
+    const tempPkg = await readJson(path.join(this.profileDir, 'package.json'));
+    const config = tempPkg['dev-server'] as DevServerConfig;
+    if (!config) {
+      throw new Error(`Couldn't find dev-server configuration in package.json`);
+    }
+    if (
+      !(await this.waitForPort(config.adminPort, OBJECTS_DB_PORT_OFFSET)) ||
+      !(await this.waitForPort(config.adminPort, STATES_DB_PORT_OFFSET))
+    ) {
+      throw new Error(`Couldn't start js-controller`);
+    }
+  }
+
   async startJsController(): Promise<void> {
-    const proc = this.spawn(
+    const proc = await this.spawn(
       'node',
       ['--inspect=127.0.0.1:9228', 'node_modules/iobroker.js-controller/controller.js'],
       this.profileDir,
     );
     proc.on('exit', async (code) => {
       console.error(chalk.yellow(`ioBroker controller exited with code ${code}`));
-      return this.exit(-1);
+      return this.exit(-1, 'SIGKILL');
     });
+    this.log.notice('Waiting for js-controller to start...');
+    await this.waitForJsController();
   }
 
   private async startJsControllerDebug(wait: boolean): Promise<void> {
@@ -506,13 +575,18 @@ class DevServer {
     } else {
       nodeArgs.unshift('--inspect');
     }
-    const proc = this.spawn('node', nodeArgs, this.profileDir);
+    const proc = await this.spawn('node', nodeArgs, this.profileDir);
     proc.on('exit', (code) => {
       console.error(chalk.yellow(`ioBroker controller exited with code ${code}`));
       return this.exit(-1);
     });
+    await this.waitForJsController();
 
     this.log.box(`Debugger is now ${wait ? 'waiting' : 'available'} on process id ${proc.pid}`);
+  }
+
+  async delay(ms: number): Promise<void> {
+    return new Promise((resolve) => setTimeout(resolve, ms));
   }
 
   async startServer(): Promise<void> {
@@ -526,12 +600,14 @@ class DevServer {
 
     const hiddenAdminPort = this.getPort(config.adminPort, HIDDEN_ADMIN_PORT_OFFSET);
 
+    await this.waitForPort(hiddenAdminPort);
+
     const app = express();
     if (this.isJSController()) {
       // simply forward admin as-is
       app.use(
         createProxyMiddleware({
-          target: `http://localhost:${hiddenAdminPort}`,
+          target: `http://127.0.0.1:${hiddenAdminPort}`,
           ws: true,
         }),
       );
@@ -548,7 +624,7 @@ class DevServer {
     const server = app.listen(config.adminPort);
 
     let exiting = false;
-    process.on('SIGINT', () => {
+    process.on('SIGINT', async () => {
       this.log.notice('dev-server is exiting...');
       exiting = true;
       server.close();
@@ -565,7 +641,7 @@ class DevServer {
       const connectWebSocketClient = (): void => {
         if (exiting) return;
         // TODO: replace this with @iobroker/socket-client
-        this.websocket = new WebSocket(`ws://localhost:${hiddenAdminPort}/?sid=${Date.now()}&name=admin`);
+        this.websocket = new WebSocket(`ws://127.0.0.1:${hiddenAdminPort}/?sid=${Date.now()}&name=admin`);
         this.websocket.on('open', () => this.log.silly('WebSocket open'));
         this.websocket.on('close', () => {
           this.log.silly('WebSocket closed');
@@ -574,9 +650,10 @@ class DevServer {
         });
         this.websocket.on('error', (error) => this.log.silly(`WebSocket error: ${error}`));
         this.websocket.on('message', (msg) => {
-          if (typeof msg === 'string') {
+          const msgString = msg && typeof msg !== 'string' ? msg.toString() : null;
+          if (typeof msgString === 'string') {
             try {
-              const data = JSON.parse(msg);
+              const data = JSON.parse(msgString);
               if (!Array.isArray(data) || data.length === 0) return;
               switch (data[0]) {
                 case 0:
@@ -599,7 +676,7 @@ class DevServer {
       connectWebSocketClient();
     }
 
-    this.log.box(`Admin is now reachable under http://localhost:${config.adminPort}/`);
+    this.log.box(`Admin is now reachable under http://127.0.0.1:${config.adminPort}/`);
   }
 
   private async createJsonConfigProxy(app: Application, config: DevServerConfig): Promise<void> {
@@ -623,7 +700,7 @@ class DevServer {
     });
 
     // "proxy" for the main page which injects our script
-    const adminUrl = `http://localhost:${this.getPort(config.adminPort, HIDDEN_ADMIN_PORT_OFFSET)}`;
+    const adminUrl = `http://127.0.0.1:${this.getPort(config.adminPort, HIDDEN_ADMIN_PORT_OFFSET)}`;
     app.get('/', async (_req, res) => {
       const { data } = await axios.get<string>(adminUrl);
       res.send(injectCode(data, this.adapterName));
@@ -632,7 +709,7 @@ class DevServer {
     // browser-sync proxy
     app.use(
       createProxyMiddleware(['/browser-sync/**'], {
-        target: `http://localhost:${browserSyncPort}`,
+        target: `http://127.0.0.1:${browserSyncPort}`,
         //ws: true, // can't have two web-socket connections proxying to different locations
       }),
     );
@@ -680,7 +757,7 @@ class DevServer {
     pathRewrite[`^/adapter/${this.adapterName}/`] = '/';
     app.use(
       createProxyMiddleware([adminPattern, '/browser-sync/**'], {
-        target: `http://localhost:${browserSyncPort}`,
+        target: `http://127.0.0.1:${browserSyncPort}`,
         //ws: true, // can't have two web-socket connections proxying to different locations
         pathRewrite,
       }),
@@ -689,7 +766,7 @@ class DevServer {
     // admin proxy
     app.use(
       createProxyMiddleware([`!${adminPattern}`, '!/browser-sync/**'], {
-        target: `http://localhost:${this.getPort(config.adminPort, HIDDEN_ADMIN_PORT_OFFSET)}`,
+        target: `http://127.0.0.1:${this.getPort(config.adminPort, HIDDEN_ADMIN_PORT_OFFSET)}`,
         ws: true,
       }),
     );
@@ -721,7 +798,7 @@ class DevServer {
       sourcemaps.map(async (sourcemap) => {
         const src = path.join(this.rootDir, sourcemap);
         const dest = path.join(outDir, sourcemap);
-        this.patchSourcemap(src, dest);
+        await this.patchSourcemap(src, dest);
       }),
     );
   }
@@ -741,7 +818,7 @@ class DevServer {
       // append the sourcemap reference comment to the bottom of the file
       const fileContent = await readFile(copyFromSrc ? src : dest, { encoding: 'utf-8' });
       const filename = path.basename(mapFile);
-      let updatedContent = fileContent.replace(/(\/\/\# sourceMappingURL=).+/, `$1${filename}`);
+      let updatedContent = fileContent.replace(/(\/\/# sourceMappingURL=).+/, `$1${filename}`);
       if (updatedContent === fileContent) {
         // no existing source mapping URL was found in the file
         if (!fileContent.endsWith('\n')) {
@@ -871,7 +948,7 @@ class DevServer {
     if (wait) {
       args.push('--wait');
     }
-    const proc = this.spawn('node', args, this.profileDir);
+    const proc = await this.spawn('node', args, this.profileDir);
     proc.on('exit', (code) => {
       console.error(chalk.yellow(`Adapter debugging exited with code ${code}`));
       return this.exit(-1);
@@ -917,7 +994,7 @@ class DevServer {
     );
   }
 
-  private async startAdapterWatch(startAdapter: boolean): Promise<void> {
+  private async startAdapterWatch(startAdapter: boolean, doNotWatch: string[]): Promise<void> {
     // figure out if we need to watch for TypeScript changes
     const pkg = await this.readPackageJson();
     const scripts = pkg.scripts;
@@ -931,7 +1008,8 @@ class DevServer {
     await this.startFileSync(adapterRunDir);
 
     if (startAdapter) {
-      await this.startNodemon(adapterRunDir, pkg.main);
+      await this.delay(3000);
+      await this.startNodemon(adapterRunDir, pkg.main, doNotWatch);
     } else {
       this.log.box(
         `You can now start the adapter manually by running\n    ` +
@@ -956,9 +1034,12 @@ class DevServer {
       const ignoreFiles = [] as string[];
       const watcher = chokidar.watch(patterns, { cwd: this.rootDir });
       let ready = false;
+      let initialEventPromises: Promise<void>[] = [];
       watcher.on('error', reject);
-      watcher.on('ready', () => {
+      watcher.on('ready', async () => {
         ready = true;
+        await Promise.all(initialEventPromises);
+        initialEventPromises = [];
         resolve();
       });
       /*watcher.on('all', (event, path) => {
@@ -988,12 +1069,15 @@ class DevServer {
           // ignore files during initial sync if they don't exist in the target directory (except for sourcemaps)
           ignoreFiles.push(filename);
         } else {
-          syncFile(filename);
+          initialEventPromises.push(syncFile(filename));
         }
       });
       watcher.on('change', (filename: string) => {
         if (!ignoreFiles.includes(filename)) {
-          syncFile(filename);
+          const resPromise = syncFile(filename);
+          if (!ready) {
+            initialEventPromises.push(resPromise);
+          }
         }
       });
       watcher.on('unlink', (filename: string) => {
@@ -1006,7 +1090,7 @@ class DevServer {
     });
   }
 
-  private async startNodemon(baseDir: string, scriptName: string): Promise<void> {
+  private async startNodemon(baseDir: string, scriptName: string, doNotWatch: string[]): Promise<void> {
     const script = path.resolve(baseDir, scriptName);
     this.log.notice(`Starting nodemon for ${script}`);
 
@@ -1017,6 +1101,11 @@ class DevServer {
 
     const args = this.isJSController() ? [] : ['--debug', '0'];
 
+    const ignoreList = [path.join(baseDir, 'admin')];
+    if (doNotWatch.length > 0) {
+      doNotWatch.forEach((entry) => ignoreList.push(path.join(baseDir, entry)));
+    }
+
     nodemon({
       script: script,
       stdin: false,
@@ -1024,7 +1113,7 @@ class DevServer {
       // dump: true, // this will output the entire config and not do anything
       colours: false,
       watch: [baseDir],
-      ignore: [path.join(baseDir, 'admin')],
+      ignore: ignoreList,
       ignoreRoot: [],
       delay: 2000,
       execMap: { js: 'node --inspect' },
@@ -1247,6 +1336,14 @@ class DevServer {
       systemConfig.common.licenseConfirmed = true; // Disable license confirmation
       systemConfig.common.defaultLogLevel = 'debug'; // Set default log level for adapters to debug
       systemConfig.common.activeRepo = 'beta'; // Set adapter repository to beta
+      // Set other details to dummy values that they are not empty like in a normal installation
+      systemConfig.common.city = 'Berlin';
+      systemConfig.common.country = 'Germany';
+      systemConfig.common.longitude = 13.28;
+      systemConfig.common.latitude = 52.5;
+      systemConfig.common.language = 'en';
+      systemConfig.common.tempUnit = '°C';
+      systemConfig.common.currency = '€';
       return systemConfig;
     });
   }
@@ -1447,29 +1544,46 @@ class DevServer {
     });
   }
 
-  private spawn(command: string, args: ReadonlyArray<string>, cwd: string, options?: cp.SpawnOptions): cp.ChildProcess {
-    this.log.debug(`${cwd}> ${command} ${args.join(' ')}`);
-    const proc = cp.spawn(command, args, {
-      stdio: ['ignore', 'inherit', 'inherit'],
-      cwd: cwd,
-      ...options,
+  private spawn(
+    command: string,
+    args: ReadonlyArray<string>,
+    cwd: string,
+    options?: cp.SpawnOptions,
+  ): Promise<cp.ChildProcess> {
+    return new Promise((resolve, reject) => {
+      let processSpawned = false;
+      this.log.debug(`${cwd}> ${command} ${args.join(' ')}`);
+      const proc = cp.spawn(command, args, {
+        stdio: ['ignore', 'inherit', 'inherit'],
+        cwd: cwd,
+        ...options,
+      });
+      this.childProcesses.push(proc);
+      let alive = true;
+      proc.on('spawn', () => {
+        processSpawned = true;
+        resolve(proc);
+      });
+      proc.on('error', (err) => {
+        this.log.error(`Could not spawn ${command}: ${err}`);
+        if (!processSpawned) {
+          reject(err);
+        }
+      });
+      proc.on('exit', () => (alive = false));
+      process.on('exit', () => alive && proc.kill('SIGINT'));
     });
-    this.childProcesses.push(proc);
-    let alive = true;
-    proc.on('exit', () => (alive = false));
-    process.on('exit', () => alive && proc.kill());
-    return proc;
   }
 
-  private spawnAndAwaitOutput(
+  private async spawnAndAwaitOutput(
     command: string,
     args: ReadonlyArray<string>,
     cwd: string,
     awaitMsg: string | RegExp,
     options?: cp.SpawnOptions,
   ): Promise<cp.ChildProcess> {
+    const proc = await this.spawn(command, args, cwd, { ...options, stdio: ['ignore', 'pipe', 'pipe'] });
     return new Promise<cp.ChildProcess>((resolve, reject) => {
-      const proc = this.spawn(command, args, cwd, { ...options, stdio: ['ignore', 'pipe', 'pipe'] });
       const handleStream = (isStderr: boolean) => (data: Buffer) => {
         let str = data.toString('utf-8');
         str = str.replace(/\x1Bc/, ''); // filter the "clear screen" ANSI code (used by tsc)
@@ -1497,7 +1611,7 @@ class DevServer {
 
       proc.on('exit', (code) => reject(`Exited with ${code}`));
       process.on('SIGINT', () => {
-        proc.kill();
+        proc.kill('SIGINT');
         reject('SIGINT');
       });
     });
@@ -1513,22 +1627,29 @@ class DevServer {
     return value.replace(/[|\\{}()[\]^$+*?.]/g, '\\$&').replace(/-/g, '\\x2d');
   }
 
-  private async exit(exitCode: number): Promise<never> {
+  private async exit(exitCode: number, signal = 'SIGINT'): Promise<never> {
     const childPids = this.childProcesses.map((p) => p.pid).filter((p) => !!p) as number[];
-    const tryKill = (pid: number): void => {
+    const tryKill = (pid: number, signal: string): void => {
       try {
-        process.kill(pid, 'SIGKILL');
+        process.kill(pid, signal);
       } catch {
         // ignore
       }
     };
     try {
       const children = await Promise.all(childPids.map((pid) => this.getChildProcesses(pid)));
-      children.forEach((ch) => ch.forEach((c) => tryKill(parseInt(c.PID))));
+      children.forEach((ch) => ch.forEach((c) => tryKill(parseInt(c.PID), signal)));
     } catch (error) {
       this.log.error(`Couldn't kill grand-child processes: ${error}`);
     }
-    childPids.forEach((pid) => tryKill(pid));
+    if (childPids.length) {
+      childPids.forEach((pid) => tryKill(pid, signal));
+      if (signal !== 'SIGKILL') {
+        // first try SIGINT and give it 5s to exit itself before killing the processes left
+        await this.delay(5000);
+        return this.exit(exitCode, 'SIGKILL');
+      }
+    }
     process.exit(exitCode);
   }
 }