+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 	## __WORK IN PROGRESS__
 -->
 
+### __WORK IN PROGRESS__
+- (Apollon77) Add option --doNotWatch for "watch" mode to ignore changes on defined files or locations
+- (Apollon77/kleinOr) Set several other system settings on "setup" call like the location, currency, language and temperature unit that they are not empty
+- (Apollon77) Only send SIGINT on exit first and then SIGKILL after 5 seconds if not yet exited to allow clean shutdown of all components
+- (Apollon77) Make sure initial file sync is finished before starting adapter (incl. a short static delay to make sure all is persisted before start watching)
+- (Apollon77) Wait that js-controller DB ports are available before starting adapter
+- (Apollon77) User 127.0.0.1 instead of localhost to avoid DNS lookup issues with Node.js 18 (looks up IPv6 by default)
+
 ## 0.5.0 (2022-04-29)
 
 - (UncleSamSwiss) Added support for JSON config UI (#164)